@@ -9,15 +9,18 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@expo/vector-icons": "^6.3.1",
+    "@expo/vector-icons": "^10.0.0",
     "axios": "^0.18.0",
-    "expo": "^30.0.1",
+    "expo": "^32.0.0",
     "jwt-decode": "^2.2.0",
-    "mobx": "^5.7.0",
-    "mobx-react": "^5.4.2",
-    "native-base": "^2.8.0",
-    "react": "16.3.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz",
-    "react-navigation": "^2.17.0"
+    "mobx": "^5.9.0",
+    "mobx-react": "^5.4.3",
+    "native-base": "^2.12.1",
+    "react": "16.5.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
+    "react-navigation": "^3.3.2"
+  },
+  "devDependencies": {
+    "babel-preset-expo": "^5.0.0"
   }
 }
