@@ -1,30 +1,25 @@
-import React from "react";
-import Expo, { AppLoading } from "expo";
-
-// Component
+import React, { Component } from "react";
+import { Root, Spinner } from "native-base";
+import { Font } from "expo";
 import HomePage from "./Components/HomePage";
 
-class App extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      fontsAreLoaded: false
-    };
-  }
+export default class App extends Component {
+  state = {
+    loading: true
+  };
 
-  componentWillMount() {
-    Expo.Font.loadAsync({
+  async componentDidMount() {
+    await Expo.Font.loadAsync({
       Roboto: require("native-base/Fonts/Roboto.ttf"),
       Roboto_medium: require("native-base/Fonts/Roboto_medium.ttf")
-    }).then(() => this.setState({ fontsAreLoaded: true }));
+    });
+    this.setState({ loading: false });
   }
 
   render() {
-    if (!this.state.fontsAreLoaded) {
-      return <AppLoading />;
+    if (this.state.loading) {
+      return <Spinner color="white" />;
     }
     return <HomePage />;
   }
 }
-
-export default App;
