@@ -19,31 +19,28 @@ import {
 import styles from "./styles";
 
 //Store
-import CoffeeStore from "../../store/coffeeStore";
+import coffeeStore from "../../store/coffeeStore";
 
 class CoffeeDetail extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      drink: "Coffee",
-      option: "Small"
-    };
-  }
+  state = {
+    drink: "Cappuccino",
+    option: "Small"
+  };
 
-  changeDrink(value) {
+  changeDrink = value => {
     this.setState({
       drink: value
     });
-  }
+  };
 
-  changeOption(value) {
+  changeOption = value => {
     this.setState({
       option: value
     });
-  }
+  };
 
   render() {
-    const coffeeshops = CoffeeStore.coffeeshops;
+    const coffeeshops = coffeeStore.coffeeshops;
     if (!coffeeshops) return <Content />;
     const coffeeshop = coffeeshops[0];
     return (
@@ -68,10 +65,10 @@ class CoffeeDetail extends Component {
                 mode="dropdown"
                 style={{ width: 150 }}
                 selectedValue={this.state.drink}
-                onValueChange={this.changeDrink.bind(this)}
+                onValueChange={this.changeDrink}
               >
-                <Picker.Item label="Coffee" value="Coffee" />
-                <Picker.Item label="Lattee" value="Lattee" />
+                <Picker.Item label="Cappuccino" value="Cappuccino" />
+                <Picker.Item label="Latte" value="Latte" />
                 <Picker.Item label="Espresso" value="Espresso" />
               </Picker>
             </Left>
@@ -81,7 +78,7 @@ class CoffeeDetail extends Component {
                 mode="dropdown"
                 style={{ width: 150 }}
                 selectedValue={this.state.option}
-                onValueChange={this.changeOption.bind(this)}
+                onValueChange={this.changeOption}
               >
                 <Picker.Item label="Small" value="Small" />
                 <Picker.Item label="Medium" value="Medium" />
