@@ -1,23 +1,21 @@
 import { decorate, observable } from "mobx";
 import axios from "axios";
-class CoffeeStore {
-  constructor() {
-    this.coffeeshops = null;
-    this.coffeeshop = null;
-    this.loading = true;
-    this.fetchAllCoffeeShops();
-  }
 
-  fetchAllCoffeeShops() {
-    axios
-      .get("http://coffee.q8fawazo.me/api/?format=json")
-      .then(res => res.data)
-      .then(coffeeshops => {
-        this.coffeeshops = coffeeshops;
-        this.loading = false;
-      })
-      .catch(err => console.error(err));
-  }
+class CoffeeStore {
+  coffeeshops = null;
+  coffeeshop = null;
+  loading = true;
+
+  fetchAllCoffeeShops = async () => {
+    try {
+      let res = await axios.get("http://coffee.q8fawazo.me/api/?format=json");
+      let coffeeshops = res.data;
+      this.coffeeshops = coffeeshops;
+      this.loading = false;
+    } catch (err) {
+      console.error(err);
+    }
+  };
 }
 
 decorate(CoffeeStore, {
@@ -26,4 +24,7 @@ decorate(CoffeeStore, {
   loading: observable
 });
 
-export default new CoffeeStore();
+let coffeeStore = new CoffeeStore();
+coffeeStore.fetchAllCoffeeShops();
+
+export default coffeeStore;
